+++ spec/spec_helper.rb
@@ -2,4 +2,17 @@ gem 'rspec', '~> 2.4'
 require 'rspec'
 require 'bundler/audit/version'
 
+module Helpers
+  def sh(command, options={})
+    Bundler.with_clean_env do
+      result = `#{command} 2>&1`
+      raise "FAILED #{command}\n#{result}" if $?.success? == !!options[:fail]
+      result
+    end
+  end
+
+  def decolorize(string)
+    string.gsub(/\e\[\d+m/, "")
+  end
+end
 include Bundler::Audit
